+++ pytsite/__init__.py
@@ -1,6 +1,5 @@
 """PytSite Init
 """
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'