+++ pytsite/router/__init__.py
@@ -5,9 +5,9 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 from ._api import handle, add_path_alias, base_path, base_url, call, current_path, current_url, dispatch, rule_path, \
-    rule_url, is_base_url, has_rule, get_no_cache, set_no_cache, remove_path_alias, scheme, server_name, url, \
-    session, request, set_request, get_session_store, on_pre_dispatch, on_dispatch, on_response, on_exception, \
-    on_xhr_dispatch, on_xhr_pre_dispatch, on_xhr_response, delete_session
+    rule_url, is_base_url, has_rule, no_cache, remove_path_alias, scheme, server_name, url, session, request, \
+    set_request, get_session_store, on_pre_dispatch, on_dispatch, on_response, on_exception, on_xhr_dispatch, \
+    on_xhr_pre_dispatch, on_xhr_response, delete_session
 
 
 def _init():